chore(UI/React): rename react components folder
